refactor: split codegen module emission
